+++ docs/README.md
@@ -5,12 +5,9 @@
 - [x] 이름 생성
     - [x] null 이거나 empty일 수 없다.
 
-### 색상
-
-- [x] 색상 생성
-    - [x] 색상은 null 이거나 empty일 수 없다.
-
 ### 구간
+
+#### 구간
 
 - [ ] 구간 생성
     - [x] 라인을 갖는다.
@@ -18,14 +15,17 @@
     - [x] 하행역을 갖는다.
     - [x] 거리를 갖는다.
 
-### 구간들
+#### 구간들
 
 - [ ] 구간들 생성
     - [ ] 구간을 추가한다.
 - [x] 역 목록을 조회한다.
 
 ### 노선
+- 
 
+- [x] 색상 생성
+    - [x] 색상은 null 이거나 empty일 수 없다.
 - [x] 지하철 노선 생성
     - [x] 이름을 갖는다.
     - [x] 색상을 갖는다.
