+++ src/main/java/nextstep/subway/line/domain/Color.java
@@ -1,4 +1,4 @@
-package nextstep.subway.common.vo;
+package nextstep.subway.line.domain;
 
 import javax.persistence.Access;
 import javax.persistence.AccessType;
