+++ src/main/java/nextstep/subway/line/domain/Section.java
@@ -1,6 +1,5 @@
 package nextstep.subway.line.domain;
 
-import nextstep.subway.common.vo.Distance;
 import nextstep.subway.station.domain.Station;
 
 import javax.persistence.*;
